toolchian: custom the package name
This release the meta-imx start to provide arm toolchain with the
same package name as the one in rtos-industrial layer. As its layer
priroity '8' is higher than rtos-industrial's priority '6', the
meta-imx provided toolchain is the effective one.

This patch is to custom rtos-industrial provided arm toolchain package
name, so that use our toolchain to build the RTOS examples.

Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>